requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.33.0.post1
torch==1.12.1
torchvision==0.13.1
pillow==9.5.0
gdown==5.2.0